pyproject.toml
[project]
name = "scriptindicacoesbazico"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gspread>=6.2.1",
    "gspread-dataframe>=4.0.0",
    "gspread-formatting>=1.2.1",
    "ipykernel>=6.29.5",
    "oauth2client>=4.1.3",
    "pandas>=2.3.1",
    "python-dotenv>=1.0.0",
]
